update ajuna cli
