+++ cli/src/trusted_commands.rs
@@ -23,7 +23,7 @@ use crate::{
 	Cli,
 };
 use codec::Decode;
-use ita_stf::{Index, KeyPair, TrustedCall, TrustedGetter, TrustedOperation};
+use ita_stf::{Index, KeyPair, SgxBoardStruct, TrustedCall, TrustedGetter, TrustedOperation};
 use log::*;
 use my_node_runtime::Balance;
 use sp_application_crypto::{ed25519, sr25519};
@@ -119,6 +119,21 @@ pub enum TrustedCommands {
 		/// amount to be transferred
 		amount: Balance,
 	},
+
+	/// Play a turn on connect four.
+	PlayTurn {
+		/// Players's incognito AccountId in ss58check format.
+		player: String,
+
+		/// Column to be played.
+		column: u8,
+	},
+
+	/// Gets the board state the given player is playing on.
+	GetBoard {
+		/// Players's incognito AccountId in ss58check format.
+		player: String,
+	},
 }
 
 pub fn match_trusted_commands(cli: &Cli, trusted_args: &TrustedArgs) {
@@ -132,6 +147,9 @@ pub fn match_trusted_commands(cli: &Cli, trusted_args: &TrustedArgs) {
 		TrustedCommands::Balance { account } => balance(cli, trusted_args, account),
 		TrustedCommands::UnshieldFunds { from, to, amount } =>
 			unshield_funds(cli, trusted_args, from, to, amount),
+		TrustedCommands::PlayTurn { player, column } =>
+			play_turn(cli, trusted_args, player, *column),
+		TrustedCommands::GetBoard { player } => get_board(cli, trusted_args, player),
 	}
 }
 
@@ -240,4 +258,59 @@ fn unshield_funds(
 			.sign(&KeyPair::Sr25519(from), nonce, &mrenclave, &shard)
 			.into_trusted_operation(trusted_args.direct);
 	let _ = perform_operation(cli, trusted_args, &top);
+}
+
+fn play_turn(cli: &Cli, trusted_args: &TrustedArgs, arg_player: &str, column: u8) {
+	let player = get_pair_from_str(trusted_args, arg_player);
+	println!("player ss58 is {}", player.public().to_ss58check());
+	println!("column choice is {:?}", column);
+
+	println!("send trusted call play-turn from {} with column {:?}", player.public(), column);
+	let (mrenclave, shard) = get_identifiers(trusted_args);
+	let nonce = get_layer_two_nonce!(player, cli, trusted_args);
+
+	let top: TrustedOperation = TrustedCall::connectfour_play_turn(
+		sr25519_core::Public::from(player.public()).into(),
+		column,
+	)
+	.sign(&KeyPair::Sr25519(player), nonce, &mrenclave, &shard)
+	.into_trusted_operation(trusted_args.direct);
+
+	let _ = perform_operation(cli, trusted_args, &top);
+}
+
+fn get_board(cli: &Cli, trusted_args: &TrustedArgs, arg_player: &str) {
+	let player = get_pair_from_str(trusted_args, arg_player);
+
+	let key_pair = sr25519_core::Pair::from(player.clone());
+
+	let top: TrustedOperation =
+		TrustedGetter::board(sr25519_core::Public::from(player.public()).into())
+			.sign(&KeyPair::Sr25519(key_pair))
+			.into();
+	let res = perform_operation(cli, trusted_args, &top);
+	debug!("received result for board");
+	if let Some(v) = res {
+		if let Ok(board) = SgxBoardStruct::decode(&mut v.as_slice()) {
+			println!("Last turn in block number: {}", board.last_turn);
+			println!("Next player: {}", board.next_player);
+			println!("Board state: {:?}", board.board_state);
+			println!("Board:");
+			for row in 0..6 {
+				for column in 0..7 {
+					print!(" {} ", board.board[column][row]);
+				}
+				println!()
+			}
+			println!("=====================");
+			for column in 0..7 {
+				print!(" {} ", column);
+			}
+			println!();
+		} else {
+			println!("could not decode board. maybe hasn't been set? {:x?}", v);
+		}
+	} else {
+		println!("could not fetch board");
+	};
 }
